# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tennis Portal API
  version: 1.0.0
  description: API для теннисного спортивного портала на FastAPI.
servers:
  - url: /api/v1
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Авторизация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          minLength: 8
        locale:
          type: string
        timezone:
          type: string
    LoginRequest:
      type: object
      required: [email_or_username, password]
      properties:
        email_or_username:
          type: string
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
        status:
          type: string
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        locale:
          type: string
        timezone:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            user:
              $ref: "#/components/schemas/User"
        meta:
          type: object
        errors:
          type: "null"
...